backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
passlib[bcrypt]
httpx
python-multipart
